# apps/hub_api/openapi.json
{
  "components": {
    "schemas": {
      "AppleSignInRequest": {
        "properties": {
          "id_token": {
            "minLength": 1,
            "title": "Id Token",
            "type": "string"
          }
        },
        "required": [
          "id_token"
        ],
        "title": "AppleSignInRequest",
        "type": "object"
      },
      "AppleSignInResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "default": 3600,
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserModel"
          }
        },
        "required": [
          "access_token",
          "user"
        ],
        "title": "AppleSignInResponse",
        "type": "object"
      },
...
      "AuthUserModel": {
        "properties": {
          "auth_provider": {
            "enum": [
              "local",
              "google",
              "apple"
            ],
            "title": "Auth Provider",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "email_verified",
          "auth_provider"
        ],
        "title": "AuthUserModel",
        "type": "object"
      },
...
      "DeviceRegistryDeleteResponse": {
        "properties": {
          "purged": {
            "title": "Purged",
            "type": "boolean"
          },
          "removed": {
            "title": "Removed",
            "type": "boolean"
          }
        },
        "required": [
          "removed",
          "purged"
        ],
        "title": "DeviceRegistryDeleteResponse",
        "type": "object"
      },
      "DeviceRegistryEntryModel": {
        "properties": {
          "addedAt": {
            "title": "Addedat",
            "type": "string"
          },
          "potId": {
            "title": "Potid",
            "type": "string"
          }
        },
        "required": [
          "potId",
          "addedAt"
        ],
        "title": "DeviceRegistryEntryModel",
        "type": "object"
      },
      "DeviceRegistryRequest": {
        "properties": {
          "potId": {
            "description": "Pot identifier to track manually.",
            "title": "Potid",
            "type": "string"
          }
        },
        "required": [
          "potId"
        ],
        "title": "DeviceRegistryRequest",
        "type": "object"
      },
      "DeviceRegistryResponse": {
        "properties": {
          "devices": {
            "items": {
              "$ref": "#/components/schemas/DeviceRegistryEntryModel"
            },
            "title": "Devices",
            "type": "array"
          }
        },
        "required": [
          "devices"
        ],
        "title": "DeviceRegistryResponse",
        "type": "object"
      },
      "DeviceRegistryUpdateResponse": {
        "properties": {
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceRegistryEntryModel"
          }
        },
        "required": [
          "device",
          "created"
        ],
        "title": "DeviceRegistryUpdateResponse",
        "type": "object"
      },
...
      "GoogleSignInRequest": {
        "properties": {
          "id_token": {
            "minLength": 1,
            "title": "Id Token",
            "type": "string"
          }
        },
        "required": [
          "id_token"
        ],
        "title": "GoogleSignInRequest",
        "type": "object"
      },
      "GoogleSignInResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "default": 3600,
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserModel"
          }
        },
        "required": [
          "access_token",
          "user"
        ],
        "title": "GoogleSignInResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocalSignInRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LocalSignInRequest",
        "type": "object"
      },
      "LocalSignInResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "challenge_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge Id"
          },
          "effective_access": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Access"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "expires_in": {
            "default": 3600,
            "title": "Expires In",
            "type": "integer"
          },
          "factor_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Factor Type"
          },
          "mfa_required": {
            "default": false,
            "title": "Mfa Required",
            "type": "boolean"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthUserModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "access_token"
        ],
        "title": "LocalSignInResponse",
        "type": "object"
      },
      "MfaVerifyRequest": {
        "properties": {
          "challenge_id": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Challenge Id",
            "type": "string"
          },
          "code": {
            "maxLength": 64,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "code"
        ],
        "title": "MfaVerifyRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "default": 3600,
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ProjectPlant Hub",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/apple": {
      "post": {
        "operationId": "sign_in_with_apple_api_v1_auth_apple_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppleSignInResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign In With Apple",
        "tags": [
          "v1",
          "auth"
        ]
      }
    },
    "/api/v1/auth/google": {
      "post": {
        "operationId": "sign_in_with_google_api_v1_auth_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleSignInResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign In With Google",
        "tags": [
          "v1",
          "auth"
        ]
      }
    },
    "/api/v1/auth/local": {
      "post": {
        "operationId": "sign_in_with_local_account_api_v1_auth_local_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalSignInResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign In With Local Account",
        "tags": [
          "v1",
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Logout Api V1 Auth Logout Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "v1",
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "operationId": "verify_mfa_api_v1_auth_mfa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalSignInResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Mfa",
        "tags": [
          "v1",
          "auth"
        ]
      }
    },
    "/api/v1/auth/token": {
      "post": {
        "operationId": "issue_token_api_v1_auth_token_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Token",
        "tags": [
          "v1",
          "auth"
        ]
      }
    },
    "/api/v1/devices": {
      "get": {
        "operationId": "list_devices_api_v1_devices_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRegistryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Devices",
        "tags": [
          "v1",
          "devices"
        ]
      },
      "post": {
        "operationId": "add_device_api_v1_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegistryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRegistryUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Device",
        "tags": [
          "v1",
          "devices"
        ]
      }
    },
    "/api/v1/devices/{pot_id}": {
      "delete": {
        "operationId": "delete_device_api_v1_devices__pot_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "pot_id",
            "required": true,
            "schema": {
              "title": "Pot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "purgeCache",
            "required": false,
            "schema": {
              "default": true,
              "title": "Purgecache",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRegistryDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Device",
        "tags": [
          "v1",
          "devices"
        ]
      }
    },
...
  }
}